requirements/development.txt
#Base
pytz==2020.1
python-slugify==4.0.1
Pillow==7.2.0
mysqlclient==2.0.1

#Django
Django==3.1.1

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.11.1
django-filter==2.3.0

# PyJWT
PyJWT==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files in divers PaaS and not ngix dependency
whitenoise==5.2.0

# Debugging
ipdb==0.13.3

# Tools
django-extensions==3.0.8

# Testing
mypy==0.782
pytest==6.0.1
pytest-sugar==0.9.4
pytest-django==3.9.0
factory-boy==3.0.1

# Code quality
flake8==3.8.3
